Connected Figure_Settings
Captain's Log, Stardate 43125.8.
Things work... tentatively. I can foresee some bugs popping up with datetime parsing, and I haven't implemented the manual datetime parse override. No input control, either - the program can't deal with arbitrary start/end datetime inputs.
With the implementation of the Figure_Settings QDockWidget, replotting the entire figure is unavoidable because I have to pass the inputs to a new GridSpec (at least I think so. I didn't look very far into 'updating' the gridspec, but if it's possible, the code is much more complicated). As it is, Matplotlib starts to slow down significantly when the subplots are populated with long dataframe series. VisPy is a faster, GPU accelerated library, but it might be underdeveloped and would probably require an enormous overhaul of the entire program to take advantage of.
Grid toggles and unit type color pickers are still disconnected, but these are minor features. More pressing is the manual format override option and the merging of the Configure and Dataframes tabs into one.
Save (Ctrl+S) works, but only for JPGs. Bare-bones.